+++ selectors_solana.yml
@@ -1,0 +1,11 @@
+selectors:
+  "5eykt4UsFv8P8NJdTREpY1vzqKqZKvdpKuc147dw2N9d":
+    name: solana-mainnet
+    selector: 124615329519749607
+  "4uhcVJyU9pJkvQyS88uRDiswHXSCkY3zQawwpjk2NsNY":
+    name: solana-testnet
+    selector: 6302590918974934319
+  "EtWTRABZaYq6iMfeYKouRu166VU2xqa1wcaWoxPkrZBG":
+    name: solana-devnet
+    selector: 16423721717087811551
+
