+++ test_selectors.yml
@@ -1,3 +1,5 @@
+# EVM test selectors.
+# File doesn't have a family suffix for backwards compatibility.
 selectors:
   # For testing purposes
   1000:
